refactor: streamline deck creation logic and remove comments
Refactor the deck creation logic by removing unnecessary comments and 
simplifying the outline generation process. Eliminate the default prompt 
selection and redundant rules, enhancing code clarity and maintainability. 
This improves the overall efficiency of the deck creation workflow.
